requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
python-multipart==0.0.6
python-dotenv==1.0.0
GitPython==3.1.40
PyYAML==6.0.2
Jinja2==3.1.2
pydantic==1.9.0
typing-extensions>=4.5.0
Markdown==3.5.1
requests==2.31.0
plantuml==0.3.0
bs4==0.0.1
beautifulsoup4==4.12.2
pylint==3.0.2
javalang==0.13.0
pymdown-extensions==10.5
pytest==7.4.3
six==1.16.0
python-markdown-math==0.8
pygments==2.16.1
# Add other dependencies as needed 